+++ SuperSecureBank/Site.Master.cs
@@ -23,7 +23,7 @@ namespace SuperSecureBank
                     LoginInfo.Text = "Please <a href=\"Account/Login.aspx\">Log in</a>";
                 else
                 {
-                    int UserID = UserMgmt.LookupSession(Request.Cookies[Settings.Default.SessionCookieKey].Value);
+                    Int64 UserID = UserMgmt.LookupSession(Request.Cookies[Settings.Default.SessionCookieKey].Value);
                     LoginInfo.Text = "Welcome to the world of secure banking, " + UserMgmt.LookupUsername(UserID) + "!<br />" +
                                                         "<a href=\"/Logout.aspx\">Log out</a>";
                 }
